requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.22
pydantic>=2.0
numpy>=1.24
scikit-learn>=1.3.0
scikit-fuzzy>=0.4.2
sentence-transformers>=2.2.0
pandas>=2.0.0
pyserial>=3.5
# AI Agent & RAG Dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
instructor>=0.4.0
# Open-source LLM support
ollama>=0.1.0
huggingface-hub>=0.20.0
transformers>=4.36.0
